feat(context): improve shell detection
Improve the detection of the default shell by considering the operating system. If the operating system is not Windows, the `SHELL` environment variable is used. Otherwise, the `ComSpec` environment variable is checked and the shell name is extracted from it. If `ComSpec` is not set, the default shell is set to "PowerShell". This change ensures accurate detection of the default shell across different operating systems.