feat(cexec): add support for PowerShell executor
This commit adds a new file `pwsh.go` which contains the implementation of the PowerShell executor for the `cexec` package. The PowerShell executor allows executing PowerShell scripts by using the `pwsh` command. The executor also supports editing the script file with the user-defined editor, or falls back to using Notepad if no editor is defined.